CoinFlip test code
I set up a simple structure in the contract. This structure simply selects a random number between the numbers 1 and 10 when the flip function is called. If this number is even, it writes the number 0 to the state. If it is odd, it writes the number 1 to the state.

Also I created a CoinFlip.test.ts file specifically for this contract. In this file, simply tests my entire code and console logs the number written to the state as a result of the flip.